Use overlay view rather than window to support iOS 13
In iOS 13 the previous hack of returning a rootViewController from a
different window no longer works. The intent of that hack was to keep the
main window controlling the status bar style and visibility, because iOS
chooses the topmost window as the controlling window with no way for
a window to opt-out (see https://openradar.appspot.com/23677818).

I tried several things to get this to work with an overlay window on
iOS 13 but could not. So I've gone to a view-based approach where I
use the layer's zIndex to ensure it's always on top.